refactor: [#639] Tracker Checker: extract mod for UDP checks
