+++ src/console/clients/checker/checks/mod.rs
@@ -1,0 +1,3 @@
+pub mod health;
+pub mod http;
+pub mod udp;
