+++ src/console/clients/checker/mod.rs
@@ -1,4 +1,5 @@
 pub mod app;
+pub mod checks;
 pub mod config;
 pub mod console;
 pub mod logger;
